# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NER Smart Logistics & Accessibility Intelligence Platform",
    "description": "Backend API for the Smart Logistics & Accessibility Intelligence Platform for North-East India (SIH 2026 \u2014 Problem 26002, MDoNER). Supports field incident reporting, road network status, routing, fleet tracking, and logistics coordination for humanitarian operations.",
    "contact": {
      "name": "SIH 2026 Team"
    },
    "license": {
      "name": "Restricted \u2014 Pilot Use Only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reports/{report_id}/triage": {
      "post": {
        "tags": [
          "Incident Adjudication & Road Traversability Lifecycle"
        ],
        "summary": "Claim observation for review (transitions to UNDER_REVIEW)",
        "operationId": "triage_report_api_v1_reports__report_id__triage_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/review": {
      "post": {
        "tags": [
          "Incident Adjudication & Road Traversability Lifecycle"
        ],
        "summary": "Adjudicate field observation (Atomic 7-way verification or rejection)",
        "operationId": "review_report_api_v1_reports__report_id__review_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents": {
      "get": {
        "tags": [
          "Incident Adjudication & Road Traversability Lifecycle"
        ],
        "summary": "List operational incidents",
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "name": "lifecycle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentLifecycle"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by operational lifecycle",
              "title": "Lifecycle"
            },
            "description": "Filter by operational lifecycle"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response List Incidents Api V1 Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "Incident Adjudication & Road Traversability Lifecycle"
        ],
        "summary": "Get operational incident detail",
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/resolve": {
      "post": {
        "tags": [
          "Incident Adjudication & Road Traversability Lifecycle"
        ],
        "summary": "Resolve incident and safely recalculate road edge traversability",
        "operationId": "resolve_incident_api_v1_incidents__incident_id__resolve_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/merge": {
      "post": {
        "tags": [
          "Incident Adjudication & Road Traversability Lifecycle"
        ],
        "summary": "Merge duplicate incident into a primary target incident",
        "operationId": "merge_incidents_api_v1_incidents__incident_id__merge_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIncidentsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EdgeImpactSpec": {
        "properties": {
          "edge_id": {
            "type": "string",
            "format": "uuid",
            "title": "Edge Id"
          },
          "affected_direction": {
            "type": "string",
            "pattern": "^(BOTH|FORWARD|BACKWARD)$",
            "title": "Affected Direction",
            "default": "BOTH"
          },
          "is_full_closure": {
            "type": "boolean",
            "title": "Is Full Closure",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "edge_id"
        ],
        "title": "EdgeImpactSpec",
        "description": "Specific road edge closure impact specification."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentLifecycle": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "MONITORING",
          "RESOLVED"
        ],
        "title": "IncidentLifecycle",
        "description": "Operational lifecycle for confirmed incidents."
      },
      "IncidentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "primary_report_id": {
            "type": "string",
            "format": "uuid",
            "title": "Primary Report Id"
          },
          "lifecycle": {
            "type": "string",
            "title": "Lifecycle"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "reopened_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reopened Reason"
          },
          "reopened_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reopened At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "primary_report_id",
          "lifecycle",
          "severity",
          "title",
          "description",
          "created_at",
          "version"
        ],
        "title": "IncidentResponse",
        "description": "Operational incident details."
      },
      "LocationPointDTO": {
        "properties": {
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "description": "Longitude in decimal degrees"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "description": "Latitude in decimal degrees"
          },
          "accuracy_m": {
            "type": "number",
            "maximum": 5000.0,
            "exclusiveMinimum": 0.0,
            "title": "Accuracy M",
            "description": "GPS horizontal accuracy radius in meters"
          },
          "location_provider": {
            "$ref": "#/components/schemas/LocationProvider",
            "description": "Sensor origin",
            "default": "GPS_HARDWARE"
          },
          "altitude_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude M",
            "description": "Elevation in meters if available"
          }
        },
        "type": "object",
        "required": [
          "longitude",
          "latitude",
          "accuracy_m"
        ],
        "title": "LocationPointDTO",
        "description": "Geographic coordinates with accuracy and sensor source."
      },
      "LocationProvider": {
        "type": "string",
        "enum": [
          "GPS_HARDWARE",
          "NETWORK_COARSE",
          "MANUAL_MAP_PICK"
        ],
        "title": "LocationProvider",
        "description": "Sensor origin for location coordinates."
      },
...
      "MergeIncidentsRequest": {
        "properties": {
          "target_incident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Incident Id",
            "description": "Primary incident to retain"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Audit notes explaining duplication"
          }
        },
        "type": "object",
        "required": [
          "target_incident_id"
        ],
        "title": "MergeIncidentsRequest",
        "description": "Request to merge a duplicate incident into a primary target incident."
      },
...
      "RejectionReason": {
        "type": "string",
        "enum": [
          "DUPLICATE",
          "INACCURATE_LOCATION",
          "SPAM_OR_INVALID",
          "UNVERIFIABLE",
          "RESOLVED_PRIOR_TO_REVIEW",
          "OTHER"
        ],
        "title": "RejectionReason",
        "description": "Mandatory reason for report rejection."
      },
...
      "ReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reporter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reporter Id"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "jurisdiction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction Id"
          },
          "client_operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Operation Id"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "review_state": {
            "type": "string",
            "title": "Review State"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "location": {
            "$ref": "#/components/schemas/LocationPointDTO"
          },
          "candidate_edge_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Edge Id"
          },
          "candidate_bridge_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Bridge Id"
          },
          "is_provisional_caution": {
            "type": "boolean",
            "title": "Is Provisional Caution"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "rejection_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Notes"
          },
          "amendment_of_report_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amendment Of Report Id"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "media_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Media Ids"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reporter_id",
          "organization_id",
          "jurisdiction_id",
          "client_operation_id",
          "report_type",
          "severity",
          "review_state",
          "description",
          "location",
          "candidate_edge_id",
          "candidate_bridge_id",
          "is_provisional_caution",
          "observed_at",
          "received_at",
          "created_at",
          "media_ids",
          "version"
        ],
        "title": "ReportResponse",
        "description": "Full field observation response."
      },
...
      "ResolutionReason": {
        "type": "string",
        "enum": [
          "REPAIRS_COMPLETED",
          "HAZARD_CLEARED",
          "MERGED_INTO",
          "FALSE_ALARM",
          "EXPIRED",
          "OTHER"
        ],
        "title": "ResolutionReason",
        "description": "Enumerated justification for closing an operational incident."
      },
      "ResolveIncidentRequest": {
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/ResolutionReason",
            "description": "Operational resolution justification"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Optional resolution audit details"
          },
          "affected_edge_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Affected Edge Ids",
            "description": "List of road edges to safely re-evaluate for accessibility recalculation"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ResolveIncidentRequest",
        "description": "Request to mark an incident as resolved."
      },
      "ReviewDecisionKind": {
        "type": "string",
        "enum": [
          "CONFIRM_INCIDENT",
          "REJECT_REPORT",
          "REQUEST_MORE_INFO"
        ],
        "title": "ReviewDecisionKind",
        "description": "Adjudication outcomes for field observations."
      },
      "ReviewDecisionRequest": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ReviewDecisionKind",
            "description": "Adjudication outcome"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Audit justification notes"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RejectionReason"
              },
              {
                "type": "null"
              }
            ],
            "description": "Mandatory if decision is REJECT_REPORT"
          },
          "existing_incident_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing Incident Id",
            "description": "Link observation to existing open incident"
          },
          "affected_edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeImpactSpec"
            },
            "type": "array",
            "title": "Affected Edges",
            "description": "Affected corridor edges"
          },
          "incident_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Title",
            "description": "Title for newly created incident"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewDecisionRequest",
        "description": "Verifier adjudication submission."
      },
      "ReviewDecisionResponse": {
        "properties": {
          "report_id": {
            "type": "string",
            "format": "uuid",
            "title": "Report Id"
          },
          "review_state": {
            "type": "string",
            "title": "Review State"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "review_state",
          "version",
          "decision"
        ],
        "title": "ReviewDecisionResponse",
        "description": "Response payload for completed review decision."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}